+++ src/common/providers/reference_resolver.py
@@ -3,7 +3,10 @@ from typing import Callable
 from common.address import Address
 from common.entity.is_reference import is_reference
 from common.exceptions import ApplicationException
-from common.providers.address_resolver import ResolvedAddress, resolve_address
+from common.providers.address_resolver.address_resolver import (
+    ResolvedAddress,
+    resolve_address,
+)
 from storage.data_source_class import DataSource
