+++ src/tests/unit/common/providers/resolve_address/test_resolve_address.py
@@ -6,7 +6,7 @@ import pytest
 
 from common.address import Address
 from common.exceptions import ApplicationException, NotFoundException
-from common.providers.address_resolver import resolve_address
+from common.providers.address_resolver.address_resolver import resolve_address
 from enums import REFERENCE_TYPES, SIMOS
 from tests.unit.mock_data.mock_document_service import get_mock_document_service
